Deleting unused objects
