+++ python-sdk/indexify/executor/agent.py
@@ -27,7 +27,6 @@ from ..functions_sdk.image import ImageInformation
 from . import image_dependency_installer
 from .api_objects import ExecutorMetadata, Task
 from .downloader import DownloadedInputs, Downloader
-from .executor_tasks import DownloadGraphTask, DownloadInputTask, ExtractTask
 from .runtime_probes import ProbeInfo, RuntimeProbes
 from .task_reporter import TaskReporter
 from .task_store import CompletedTask, TaskStore
