+++ python-sdk/indexify/executor/function_worker/function_worker_utils.py
@@ -1,18 +1,32 @@
 import os
+from functools import cache
 
-import nanoid
+import cloudpickle
 
+from indexify import IndexifyClient
+from indexify.functions_sdk.indexify_functions import GraphInvocationContext, IndexifyFunctionWrapper
 
-def get_optimal_process_count():
-    """
-    Returns a reasonable number of processes based on CPU cores.
-    Generally CPU cores - 1 to leave one core for the OS/other tasks.
-    """
-    return max(os.cpu_count() - 1, 1)
-
-
-def job_generator() -> str:
-    """
-    Generates job ID
-    """
-    return nanoid.generate()
+@cache
+def _load_function(
+    namespace: str,
+    graph_name: str,
+    fn_name: str,
+    code_path: str,
+    version: int,
+    invocation_id: str,
+    indexify_client: IndexifyClient,
+):
+    """Load an extractor to the memory: extractor_wrapper_map."""
+    with open(code_path, "rb") as f:
+        code = f.read()
+        pickled_functions = cloudpickle.loads(code)
+    context = GraphInvocationContext(
+        invocation_id=invocation_id,
+        graph_name=graph_name,
+        graph_version=str(version),
+        indexify_client=indexify_client,
+    )
+    return IndexifyFunctionWrapper(
+        cloudpickle.loads(pickled_functions[fn_name]),
+        context,
+    )
